[REFACTOR] 외래키 이름 및 패키지 이름 변경 #92
