+++ src/main/java/com/sports/server/command/game/domain/GameTeam.java
@@ -1,6 +1,6 @@
 package com.sports.server.command.game.domain;
 
-import com.sports.server.command.team.domain.LeagueTeam;
+import com.sports.server.command.leagueteam.LeagueTeam;
 import com.sports.server.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -18,7 +18,7 @@ public class GameTeam extends BaseEntity<GameTeam> {
     private Game game;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "team_id")
+    @JoinColumn(name = "league_team_id")
     private LeagueTeam leagueTeam;
 
     @Column(name = "cheer_count", nullable = false)
